+++ app/src/main/java/com/steadyroom/project_d/User.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class User {
     public String user_id;
     public String nickname;
-    public int userPoint = 0;
+    public long userPoint = 0;
     public List<CharacterInstance> characters = new ArrayList<>();  // 빈 리스트로 초기화
     public List<Item> items = new ArrayList<>();
